requirements.txt
paho-mqtt==2.1.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1
typing-extensions==4.15.0
